ref(slack): update tests for bootstrap
